Adding 3B model and tokenizer and tested with the tokenizer

+++ experiments/9_training_stack_optimisation_and_cost_governor/training/deepspeed_template/src/models/model_3b.py
@@ -1012,7 +1012,7 @@ class MoEFFN(nn.Module):
 
         sort_idx = real_expert_indices.argsort()
         sorted_token_indices = real_token_indices[sort_idx]
-        sorted_weights = real_weights[sort_idx]
+        sorted_weights = real_weights[sort_idx].to(dtype)
         sorted_x = flat_x[sorted_token_indices]
 
         expert_counts = torch.bincount(real_expert_indices, minlength=E)
@@ -1402,7 +1402,7 @@ class Model3B(nn.Module):
         self.layer_types = layer_types
 
         # Reversible Midpoint Integration
-        from reversible_ops_midpoint import ReversibleMidpointStack
+        from .reversible_ops_midpoint import ReversibleMidpointStack
         self.stack = ReversibleMidpointStack(
             self.layers,
             step_size=0.25,
